Changed int varaibles to longs, created an ouput service with cerresponding interface, added a service/interface for printing results from the prime service.

+++ src/main/java/com/wynnblevins/isprime/service/InputService.java
@@ -1,5 +1,5 @@
 package com.wynnblevins.isprime.service;
 
 public interface InputService {
-	public Integer getInput(String promptString);
+	public Long getInput(String promptString);
 }
