+++ src/main/java/com/wynnblevins/isprime/service/PrimeService.java
@@ -1,5 +1,5 @@
 package com.wynnblevins.isprime.service;
 
 public interface PrimeService {
-	public boolean isPrime(int number);
+	public boolean isPrime(Long number);
 }
